backend/requirements.txt
numpy
Flask
python-dotenv
opencv-python
ffmpeg-python
sqlalchemy
opencv-python-headless
requests
flask-cors
flask-sqlalchemy